pyproject.toml
[project]
name = "mcp-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.3.1",
    "dotenv>=0.9.9",
    "fastmcp>=3.2.4",
    "isort>=8.0.1",
    "langchain>=1.2.11",
    "langchain-google-genai>=4.2.2",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.2.0",
    "pydantic>=2.13.4",
]
